feat(component / calendar): Se actualizo la UI del modal del calendar
