+++ projects/ds-ng/src/lib/components/bmb-calendar/common/bmb-calendar-schedule-cards/bmb-calendar-schedule-cards.component.html
@@ -3,12 +3,15 @@
     class="bmb_calendar-event"
     [ngClass]="getClassNames()"
     [style]="getPosition()"
-    (click)="handleSelectEvent($event)"
+    (click)="handleSelectEvent()"
     title="{{ event().title }} - {{ event().detail }}"
   >
     <div class="bmb_calendar-event-indicator"></div>
     <header class="bmb_calendar-event-header">
-      <span class="bmb_calendar-event-header-circle" [ngStyle]="getBulletStyle()"></span>
+      <span
+        class="bmb_calendar-event-header-circle"
+        [ngStyle]="getBulletStyle()"
+      ></span>
       <h3 class="bmb_calendar-event-title">
         {{ event().title }}
       </h3>
@@ -18,3 +21,86 @@
     }
   </div>
 }
+
+<ng-template #detailContent>
+  <div
+    bmbLayoutGrid
+    colGapSize="m"
+    rowGapSize="m"
+    columnSize="1fr"
+    rowSize="auto"
+    [columns]="2"
+    [rows]="4"
+  >
+    <section
+      bmbLayoutGridItem
+      [colStart]="1"
+      [rowStart]="1"
+      [numberOfColumns]="2"
+      [numberOfRows]="1"
+    >
+      @if (event().tags?.length) {
+        <p style="display: flex; flex-wrap: wrap; gap: 0.5rem; margin: 0">
+          @for (item of event().tags; track $index) {
+            <bmb-badge [text]="item.text" [appearance]="item.appearance" />
+          }
+        </p>
+      }
+    </section>
+    <section
+      bmbLayoutGridItem
+      [colStart]="1"
+      [rowStart]="2"
+      [numberOfColumns]="2"
+      [numberOfRows]="1"
+    >
+      <bmb-divider [removeMargin]="true" />
+    </section>
+    <section
+      bmbLayoutGridItem
+      [colStart]="1"
+      [rowStart]="3"
+      [numberOfColumns]="1"
+      [numberOfRows]="1"
+    >
+      <span class="bmb_calendar-event-modal-text">{{ getDuration() }}</span>
+    </section>
+    <section
+      bmbLayoutGridItem
+      [colStart]="2"
+      [rowStart]="3"
+      [numberOfColumns]="1"
+      [numberOfRows]="1"
+    >
+      @if (event().daysRepetition) {
+        <span class="bmb_calendar-event-modal-text">{{
+          event().daysRepetition
+        }}</span>
+      }
+    </section>
+
+    <section
+      bmbLayoutGridItem
+      [colStart]="1"
+      [rowStart]="4"
+      [numberOfColumns]="1"
+      [numberOfRows]="1"
+    >
+      @if (event().place) {
+        <p class="bmb_calendar-event-modal-text">{{ event().place }}</p>
+      }
+    </section>
+
+    <section
+      bmbLayoutGridItem
+      [colStart]="2"
+      [rowStart]="4"
+      [numberOfColumns]="1"
+      [numberOfRows]="1"
+    >
+      @if (event().campus) {
+        <span class="bmb_calendar-event-modal-text">{{ event().campus }}</span>
+      }
+    </section>
+  </div>
+</ng-template>
