+++ projects/ds-ng/src/lib/components/bmb-calendar/common/bmb-calendar-schedule-cards/bmb-calendar-schedule-cards.component.scss
@@ -78,7 +78,6 @@ $types: 'academic', 'life', 'event', 'save_the_date';
   }
 
   &-grid {
-
     &-reduced {
       .bmb_calendar-event-title {
         -webkit-line-clamp: 1;
@@ -144,4 +143,8 @@ $types: 'academic', 'life', 'event', 'save_the_date';
   &-time {
     grid-area: time;
   }
+}
+
+.bmb_calendar-event-modal-text {
+  @include fonts.font(4, regular);
 }
