+++ projects/ds-ng/src/lib/components/bmb-calendar/common/bmb-calendar-template-day/bmb-calendar-template-day.component.html
@@ -10,10 +10,7 @@
         <div class="bmb_calendar-row-body"></div>
       }
       @for (event of eventsWithLayout(); track $index) {
-        <bmb-calendar-schedule-cards
-          [event]="event"
-          (onSelectEvent)="handleEventSelection($event)"
-        />
+        <bmb-calendar-schedule-cards [event]="event" />
       }
     </div>
   </section>
