+++ projects/ds-ng/src/lib/components/bmb-calendar/common/bmb-calendar-template-month/bmb-calendar-template-month.component.html
@@ -15,7 +15,6 @@
       @for (event of eventsOnDate()[$index]; track $index) {
         <bmb-calendar-schedule-cards
           [event]="event"
-          (onSelectEvent)="handleEventSelection($event)"
           [isPositionAbsolute]="false"
           [extendedContent]="false"
         />
