+++ projects/ds-ng/src/lib/components/bmb-calendar/common/bmb-calendar-template-week/bmb-calendar-template-week.component.html
@@ -1,7 +1,10 @@
 <div class="bmb_calendar-week-view">
   <bmb-calendar-hour-view [startBusinessHour]="startBusinessHour()" />
   @for (day of dayNames; track $index) {
-    @if (($index && $index < 6) || ($index === 6 && eventsOnWeek()[$index].length > 0)) {
+    @if (
+      ($index && $index < 6) ||
+      ($index === 6 && eventsOnWeek()[$index].length > 0)
+    ) {
       <section
         class="bmb_calendar-col"
         [ngClass]="{ 'bmb_calendar-col-today': isNow(weekDays()[$index]) }"
@@ -18,10 +21,7 @@
           }
 
           @for (singleEvent of eventsOnWeek()[$index]; track $index) {
-            <bmb-calendar-schedule-cards
-              [event]="singleEvent"
-              (onSelectEvent)="handleEventSelection($event)"
-            />
+            <bmb-calendar-schedule-cards [event]="singleEvent" />
           }
         </div>
       </section>
