+++ projects/ds-ng/src/lib/components/bmb-external-link/bmb-external-link.component.ts
@@ -104,8 +104,6 @@ export class BmbExternalLinkComponent {
   }
 
   handleOpenMenu(): void {
-    console.log('Opening menu', this.contentRef?.nativeElement);
-
     this.projectionService.openContent({
       content: BmbDropdownContentComponent,
       targetRef: this.contentRef?.nativeElement,
