+++ projects/ds-ng/src/lib/components/bmb-tooltip/bmb-tooltip.component.ts
@@ -45,9 +45,8 @@ export class BmbTooltipComponent {
   @ViewChild('contentTooltip', { static: true })
   contentTooltip!: TemplateRef<any>;
 
-  constructor(
-    // private contentProjected: BmbProjectionContentService
-  ) {
+  constructor() // private contentProjected: BmbProjectionContentService
+  {
     effect(() => {
       const deprecatedTitle = this.title();
       const newTitle = this.componentTitle();
